pyproject.toml
[project]
name = "juliaPIV"
version = "0.0.10"
authors = [
  { name="Robin Shindelman", email="roshindelman@gmail.com" },
]
description = "A pipeline for running JuliaPIV based on the workflow developed at OSU CEOAS"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

dependencies = [
    "click>=8.0.0",
    "PyYAML>=6.0.2",
    "tqdm>=4.60.0",
    "juliacall>=0.9.25",
    "natsort>=8.4.0",
]

[project.scripts]
batcher = "juliaPIV.batcher.batcher:batcher_cli"
pivpipe = "juliaPIV.pivpipe.pivpipe:pivpipe_main_cli"
juliaPIV = "juliaPIV.juliaPIV:juliaPIV"

[tool.setuptools.package-data]
"juliaPIV" = ["PIVPipelineUtility/**/*"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/shindelr/juliaPIV"
Issues = "https://github.com/shindelr/juliaPIV/issues"